refactoring class in new package
