+++ src/main/java/org/example/c08/Abstraction/Animal.java
@@ -1,4 +1,4 @@
-package org.example.c07.Abstraction;
+package org.example.c08.Abstraction;
 
 abstract class Mammal extends Animal {
