+++ src/main/java/org/example/c08/Abstraction/Bread.java
@@ -1,4 +1,4 @@
-package org.example.c07.Abstraction;
+package org.example.c08.Abstraction;
 
 public class Bread extends ProductForSale{
     public Bread(String type, double price, String description) {
