+++ src/main/java/org/example/c08/Abstraction/Dog.java
@@ -1,4 +1,4 @@
-package org.example.c07.Abstraction;
+package org.example.c08.Abstraction;
 
 public class Dog extends Mammal {
