+++ src/main/java/org/example/c08/Abstraction/Egg.java
@@ -1,4 +1,4 @@
-package org.example.c07.Abstraction;
+package org.example.c08.Abstraction;
 
 public class Egg extends ProductForSale{
     public Egg(String type, double price, String description) {
