+++ src/main/java/org/example/c08/Abstraction/NodeList.java
@@ -1,4 +1,4 @@
-package org.example.c07.Abstraction;
+package org.example.c08.Abstraction;
 /*
 It has four methods:
         -  getRoot(), addItem(), removeItem() and traverse() which are package-private and abstract
